apply review notes
