+++ internal/statestore/backfill.go
@@ -25,7 +25,7 @@ import (
 	"open-match.dev/open-match/pkg/pb"
 )
 
-// CreateBackfill creates a new Backfill in the state storage. If the id already exists, it will be overwritten.
+// CreateBackfill creates a new Backfill in the state storage. The xids algorithm used to create the ids ensures that they are unique with no system wide synchronization. Calling clients are forbidden from choosing an id during create. So no conflicts will occur.
 func (rb *redisBackend) CreateBackfill(ctx context.Context, backfill *pb.Backfill) error {
 	redisConn, err := rb.redisPool.GetContext(ctx)
 	if err != nil {
